+++ src/components/interviewApply/ApplyInfo.style.ts
