+++ src/pages/interviewApply/EditInfo2.tsx
@@ -1,0 +1,21 @@
+import * as In from './InterviewApply.style';
+import ApplyInfoForm from '../../components/interviewApply/ApplyInfoForm';
+
+function EditInfo2() {
+
+	return (
+		<div>	
+            {/*Header*/}
+            <In.InputInfoHeader>
+            <In.Ellipse1/>
+            지원 정보 수정
+            </In.InputInfoHeader>
+            <img src={process.env.PUBLIC_URL + '/images/lineCircle4853.svg'}/>   
+
+            {/*Main(지원 정보 입력폼)*/}
+            <ApplyInfoForm/>
+		</div>
+	);
+}
+
+export default EditInfo2;
